# src/main/java/organizationmanagement/model/Organization.java
package organizationmanagement.model;

import jakarta.persistence.*;
import lombok.*;

import java.util.List;

@Entity
@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class Organization {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    private String name;

    @Enumerated(EnumType.STRING)
    private OrganizationType type;

    @OneToMany(mappedBy = "organization", cascade = CascadeType.ALL)
    private List<Department> departments;
}